fix: move api to `pytauri_wheel.lib` to avoid circular import error
